+++ vite.config.ts
@@ -3,5 +3,7 @@ import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()]
+  plugins: [react()],
+  base: "",
+  assetsInclude: ['**/*.glb']
 })
